+++ tk/src/rectobject.cpp
@@ -77,7 +77,7 @@ void RectObject::drawWorld() {
     */
 }
 
-void RectObject::drawScreen(bool drawCenterPoint, float cellsPerMeter, int pixelsPerCell, int doriScreenX, int doriScreenY) {
+void RectObject::drawScreen(bool drawCenterPoint, float cellsPerMeter, float pixelsPerCell, int doriScreenX, int doriScreenY) {
     // # cells = (cells / meter) * (worldOffset in meters)
     // # cells * pixelsPerCell = position in pixels
     float screenOffsetX = (cellsPerMeter * worldOffsetX) * pixelsPerCell;
